Remove separate upload-marketing-icons action, integrate functionality within existing upload-icons
