+++ .github/actions/upload-icons/src/upload-icons.js
@@ -1,12 +1,6 @@
-import { resolve, sep } from 'node:path';
-import { readFileSync } from 'node:fs';
 import { PutObjectCommand, S3Client } from '@aws-sdk/client-s3';
-import { getAssets } from './get-assets.js';
-import { getManifest } from './get-manifest.js';
-import { getCategories } from './get-categories.js';
-import { getMap } from './get-map.js';
-
-const CACHE_FOREVER = 'public, max-age=604800';
+import { getIconsRequests } from './requests/icons.js';
+import { getMarketingRequests } from './requests/marketing-icons.js';
 
 export async function upload(config) {
 	const s3 = new S3Client({
@@ -16,100 +10,19 @@ export async function upload(config) {
 			secretAccessKey: config.secretAccessKey,
 		},
 	});
-	const requests = [];
 
-	const sourceDirs = Array.isArray(config.sourceDirs)
-		? config.sourceDirs
-		: [config.sourceDirs];
+	const requests =
+		config.mode === 'marketing-icons'
+			? getMarketingRequests(config)
+			: getIconsRequests(config);
 
-	for (const dir of sourceDirs) {
-		const jsonString = readFileSync(resolve(dir, 'index.json'), 'utf-8');
-		const dirName = dir.split(sep).at(-3);
-		const indexName = [dirName.replace('icons', '').replace('-', ''), 'index']
-			.filter(Boolean)
-			.reverse()
-			.join('-');
-
-		requests.push({
-			Bucket: config.bucket,
-			Key: `${config.baseFolder}/v${config.version}/${indexName}.json`,
-			Body: jsonString,
-			ContentType: 'application/json',
-			CacheControl: CACHE_FOREVER,
-		});
-	}
-
-	const entries = sourceDirs.flatMap((dir) => {
-		const jsonString = readFileSync(resolve(dir, 'index.json'), 'utf-8');
-		return JSON.parse(jsonString).map((entry) => ({ ...entry, dir }));
-	});
-
-	const assets = getAssets(entries);
-
-	requests.push({
-		Bucket: config.bucket,
-		Key: `${config.baseFolder}/latest`,
-		Body: 'Redirect',
-		WebsiteRedirectLocation: `/v${config.version}/manifest.json`,
-		CacheControl: 'public, max-age=600',
-	});
-
-	requests.push({
-		Bucket: config.bucket,
-		Key: `${config.baseFolder}/v${config.version}/manifest.json`,
-		Body: JSON.stringify(getManifest(entries), null, 2),
-		ContentType: 'application/json',
-		CacheControl: CACHE_FOREVER,
-	});
-
-	requests.push({
-		Bucket: config.bucket,
-		Key: `${config.baseFolder}/v${config.version}/categories.json`,
-		Body: JSON.stringify(getCategories(entries), null, 2),
-		ContentType: 'application/json',
-		CacheControl: CACHE_FOREVER,
-	});
-
-	requests.push({
-		Bucket: config.bucket,
-		Key: `${config.baseFolder}/v${config.version}/map.json`,
-		Body: JSON.stringify(getMap(entries), null, 2),
-		ContentType: 'application/json',
-		CacheControl: CACHE_FOREVER,
-	});
-
-	for (const asset of assets) {
-		// Hashed actual file
-		requests.push({
-			Bucket: config.bucket,
-			Key: `${config.baseFolder}/${asset.hash}.svg`,
-			Body: asset.svg,
-			ContentType: 'image/svg+xml',
-			CacheControl: CACHE_FOREVER,
-		});
-
-		// Versioned icon redirect
-		requests.push({
-			Bucket: config.bucket,
-			Key: `${config.baseFolder}/v${config.version}/${asset.id}.svg`,
-			Body: 'Redirect',
-			WebsiteRedirectLocation: `/${asset.hash}.svg`,
-			CacheControl: CACHE_FOREVER,
-		});
-
-		if (!asset.aliases || asset.aliases.length === 0) continue;
-
-		for (const alias of asset.aliases) {
-			requests.push({
-				Bucket: config.bucket,
-				Key: `${config.baseFolder}/v${config.version}/${alias}.svg`,
-				Body: 'Redirect',
-				WebsiteRedirectLocation: `/${asset.hash}.svg`,
-			});
+	if (config.dry) {
+		console.log('Dry run — the following keys would be uploaded:');
+		for (const r of requests) {
+			console.log(' ', r.Key);
 		}
+		return;
 	}
-
-	if (config.dry) return;
 
 	await Promise.all(
 		requests.map((request) =>
