--- .github/actions/upload-marketing-icons/.gitignore
@@ -1,1 +1,0 @@
-!dist/
