--- .github/actions/upload-marketing-icons/action.yml
@@ -1,38 +1,0 @@
-name: vivid marketing icons CDN upload
-description: Uploads marketing icon PNG assets to Vivid's Icons AWS S3 bucket into the /marketing subfolder
-branding:
-  color: purple
-
-inputs:
-  access-key:
-    description: S3 bucket Access Key
-    required: true
-
-  secret-access-key:
-    description: S3 bucket Secret Access Key
-    required: true
-
-  version:
-    description: Marketing icons package version
-    required: true
-
-  source-dir:
-    description: Directory containing index.json and *.png files
-    required: false
-    default: ./libs/marketing-icons/src/generated/
-
-  bucket:
-    description: Target S3 bucket name
-    required: false
-
-  base-folder:
-    description: Root folder in the bucket where all assets live
-    required: false
-
-  region:
-    description: S3 bucket region
-    required: false
-
-runs:
-  using: node20
-  main: dist/index.js
