--- .github/actions/upload-marketing-icons/package.json
@@ -1,16 +1,0 @@
-{
-	"private": true,
-	"scripts": {
-		"build": "esbuild src/index.js --bundle --platform=node --target=node20 --format=cjs --outfile=dist/index.js --tree-shaking=true --minify",
-		"manual": "node manual.js"
-	},
-	"dependencies": {
-		"@actions/core": "^1.10.0",
-		"@aws-sdk/client-s3": "^3.989.0"
-	},
-	"devDependencies": {
-		"citty": "^0.2.1",
-		"esbuild": "^0.27.3"
-	},
-	"packageManager": "pnpm@^10.23.0"
-}
