+++ .github/workflows/_publish-marketing-icons.yml
@@ -47,10 +47,12 @@ jobs:
           echo "version=$VERSION" >> "$GITHUB_OUTPUT"
 
       - name: Publish marketing icons to CDN
-        uses: ./.github/actions/upload-marketing-icons
+        uses: ./.github/actions/upload-icons
         with:
+          mode: marketing-icons
           access-key: ${{ secrets.ICONS_S3_ACCESS_KEY }}
           secret-access-key: ${{ secrets.ICONS_S3_SECRET_ACCESS_KEY }}
           bucket: ${{ vars.ICONS_BUCKET }}
           base-folder: ${{ vars.ICONS_BASE_FOLDER }}
           version: ${{ steps.get-version.outputs.version }}
+          source-dir: ./libs/marketing-icons/src/generated/
